refactor: 검증 관련 패키지를 dto와 동일한 위치로 이동
