+++ src/main/java/roomescape/service/dto/request/ReservationRequest.java
@@ -3,7 +3,7 @@ package roomescape.service.dto.request;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
-import roomescape.web.exception.validation.DateFormat;
+import roomescape.service.dto.validation.DateFormat;
 
 public record ReservationRequest(
         @NotBlank String name,
