+++ src/main/java/roomescape/service/dto/request/ReservationTimeRequest.java
@@ -2,7 +2,7 @@ package roomescape.service.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import roomescape.domain.ReservationTime;
-import roomescape.web.exception.validation.TimeFormat;
+import roomescape.service.dto.validation.TimeFormat;
 
 import java.time.LocalTime;
