+++ src/main/java/roomescape/service/dto/validation/DateFormatValidator.java
@@ -1,16 +1,16 @@
-package roomescape.web.exception.validation;
+package roomescape.service.dto.validation;
 
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
 
-import java.time.LocalTime;
+import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
 
-public class TimeFormatValidator implements ConstraintValidator<TimeFormat, String> {
+public class DateFormatValidator implements ConstraintValidator<DateFormat, String> {
     @Override
     public boolean isValid(final String value, final ConstraintValidatorContext context) {
         try {
-            LocalTime.parse(value);
+            LocalDate.parse(value);
             return true;
         } catch (DateTimeParseException e) {
             return false;
