+++ src/main/java/roomescape/service/dto/validation/TimeFormatValidator.java
@@ -1,16 +1,16 @@
-package roomescape.web.exception.validation;
+package roomescape.service.dto.validation;
 
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
 
-import java.time.LocalDate;
+import java.time.LocalTime;
 import java.time.format.DateTimeParseException;
 
-public class DateFormatValidator implements ConstraintValidator<DateFormat, String> {
+public class TimeFormatValidator implements ConstraintValidator<TimeFormat, String> {
     @Override
     public boolean isValid(final String value, final ConstraintValidatorContext context) {
         try {
-            LocalDate.parse(value);
+            LocalTime.parse(value);
             return true;
         } catch (DateTimeParseException e) {
             return false;
